example/requirements.txt
coreapi==2.3.3
Django==2.0.1
django-cors-headers==2.1.0
djangorestframework==3.7.7
